+++ src/LaTeXBlocks.Word.AddIn/LaTeXBlocksRibbon.cs
@@ -38,6 +38,7 @@ namespace LaTeXBlocks.Word
                    "<button id=\"" + ReflowFrameControlId +
                    "\" label=\"Reflow Frame\" imageMso=\"RefreshAll\" getEnabled=\"GetReflowFrameEnabled\" onAction=\"OnReflowFrame\"/>" +
                    "<button id=\"LaTeXBlocks.UpdateEquationNumbers\" label=\"Update Numbers\" imageMso=\"FieldsUpdate\" onAction=\"OnUpdateEquationNumbers\"/>" +
+                   "<button id=\"LaTeXBlocks.CopyAsLaTeX\" label=\"Copy as LaTeX\" imageMso=\"Copy\" onAction=\"OnCopyAsLaTeX\"/>" +
                    "<editBox id=\"" + WidthControlId +
                    "\" label=\"Typesetting width (pt)\" sizeString=\"000.0\" getText=\"GetWidthText\" getEnabled=\"GetWidthEnabled\" onChange=\"OnWidthChanged\"/>" +
                    "</group></tab></tabs></ribbon></customUI>";
@@ -51,6 +52,7 @@ namespace LaTeXBlocks.Word
         public void OnEdit(Office.IRibbonControl control) { Run(addIn.ShowEditEditor); }
         public void OnReflowFrame(Office.IRibbonControl control) { Run(addIn.ReflowSelectedBlockFrame); }
         public void OnUpdateEquationNumbers(Office.IRibbonControl control) { Run(addIn.UpdateEquationNumbers); }
+        public void OnCopyAsLaTeX(Office.IRibbonControl control) { Run(addIn.CopySelectionAsLaTeX); }
         public bool GetReflowFrameEnabled(Office.IRibbonControl control)
         {
             try { return addIn.HasSelectedBlockFrame(); }